backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
gunicorn
psycopg2-binary
pandas
openpyxl
requests
python-dotenv
